refactor(activerecord): instantiate_records' eager arm is one JoinDependency#instantiate call
The inverse-of caching, the instantiation.active_record payload and the
per-record block move into JoinDependency#instantiate (join_dependency.rb:105-151),
where Rails does that work, so Relation#instantiate_records is Rails'
two-arm body with no intermediate helper. Retires the baselined
instantiate_records/instantiate call-mismatch row.
