+++ scripts/api-compare/call-mismatches-unreviewed/activerecord/relation.json
@@ -1,3 +1,3 @@
 {
-  "max": 10
+  "max": 9
 }
